Move custom controls to dedicated header
