+++ NeuralAmpModeler/projects/NeuralAmpModeler-app.vcxproj.filters
@@ -334,6 +334,7 @@
     <ClInclude Include="..\NeuralAmpModelerCore\NAM\wavenet.h">
       <Filter>NAM</Filter>
     </ClInclude>
+    <ClInclude Include="..\NeuralAmpModelerControls.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="resources">
