+++ NeuralAmpModeler/projects/NeuralAmpModeler-vst3.vcxproj.filters
@@ -433,6 +433,7 @@
     <ClInclude Include="..\NeuralAmpModelerCore\NAM\wavenet.h">
       <Filter>NAM</Filter>
     </ClInclude>
+    <ClInclude Include="..\NeuralAmpModelerControls.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="resources">
